+++ pennylane_lightning/core/src/simulators/lightning_gpu/gates/tests/Test_StateVectorCudaManaged_Generators.cpp
@@ -160,6 +160,42 @@ TEST_CASE("Generators::applyGeneratorPhaseShift", "[GateGenerators]") {
     }
 }
 
+TEST_CASE("Generators::applyGeneratorPSWAP", "[GateGenerators]") {
+    std::vector<typename StateVectorCudaManaged<double>::CFP_t> matrix{
+        {0.0, 0.0}, {0.0, 0.0}, {0.0, 0.0}, {0.0, 0.0}, {0.0, 0.0}, {0.0, 0.0},
+        {1.0, 0.0}, {0.0, 0.0}, {0.0, 0.0}, {1.0, 0.0}, {0.0, 0.0}, {0.0, 0.0},
+        {0.0, 0.0}, {0.0, 0.0}, {0.0, 0.0}, {0.0, 0.0},
+    };
+    std::mt19937 re{1337U};
+
+    for (std::size_t num_qubits = 1; num_qubits <= 5; num_qubits++) {
+        for (std::size_t applied_qubit = 0; applied_qubit < num_qubits;
+             applied_qubit++) {
+            auto init_state =
+                createRandomStateVectorData<double>(re, num_qubits);
+
+            StateVectorCudaManaged<double> psi(init_state.data(),
+                                               init_state.size());
+            StateVectorCudaManaged<double> psi_direct(init_state.data(),
+                                                      init_state.size());
+            StateVectorCudaManaged<double> psi_dispatch(init_state.data(),
+                                                        init_state.size());
+
+            std::string cache_gate_name = "DirectGenPSWAP" +
+                                          std::to_string(applied_qubit) + "_" +
+                                          std::to_string(num_qubits);
+
+            psi.applyGeneratorPSWAP({applied_qubit}, false);
+            psi_direct.applyOperation(cache_gate_name, {applied_qubit}, false,
+                                      {0.0}, matrix);
+            psi_dispatch.applyGenerator({"PSWAP"}, {applied_qubit}, false);
+
+            CHECK(psi.getDataVector() == psi_direct.getDataVector());
+            CHECK(psi_dispatch.getDataVector() == psi_direct.getDataVector());
+        }
+    }
+}
+
 TEST_CASE("Generators::applyGeneratorIsingXX", "[GateGenerators]") {
     // grad(IsingXX)() = e^{-i*0.5*a*(kron(X, X))}) => -0.5*i*(kron(X, X))
     std::vector<typename StateVectorCudaManaged<double>::CFP_t> matrix{
@@ -1770,7 +1806,7 @@ TEMPLATE_TEST_CASE("Generators::applyControlledGenerator - 2 wires",
     const std::string controlled_gate_name =
         GENERATE("IsingXX", "IsingXY", "IsingYY", "IsingZZ", "SingleExcitation",
                  "SingleExcitationMinus", "SingleExcitationPlus", "GlobalPhase",
-                 "MultiRZ");
+                 "MultiRZ", "PSWAP");
     const std::size_t num_qubits = 4;
     const TestType ep_deriv = 1e-3;
     const TestType ep_margin = 1e-4;