+++ pennylane_lightning/core/src/simulators/lightning_qubit/gates/AssignKernelMap_Default.cpp
@@ -120,6 +120,9 @@ void assignKernelsForGateOp_Default() {
     instance.assignKernelForOp(GateOperation::CRot, all_threading,
                                all_memory_model, all_qubit_numbers,
                                KernelType::LM);
+    instance.assignKernelForOp(GateOperation::PSWAP, all_threading,
+                               all_memory_model, all_qubit_numbers,
+                               KernelType::LM);
 
     /* Three-qubit gates */
     instance.assignKernelForOp(GateOperation::Toffoli, all_threading,
@@ -200,6 +203,9 @@ void assignKernelsForGeneratorOp_Default() {
     instance.assignKernelForOp(GeneratorOperation::ControlledPhaseShift,
                                all_threading, all_memory_model,
                                all_qubit_numbers, KernelType::LM);
+    instance.assignKernelForOp(GeneratorOperation::PSWAP, all_threading,
+                               all_memory_model, all_qubit_numbers,
+                               KernelType::LM);
 
     instance.assignKernelForOp(GeneratorOperation::SingleExcitation,
                                all_threading, all_memory_model,
@@ -281,6 +287,9 @@ void assignKernelsForControlledGateOp_Default() {
     instance.assignKernelForOp(ControlledGateOperation::IsingZZ, all_threading,
                                all_memory_model, all_qubit_numbers,
                                KernelType::LM);
+    instance.assignKernelForOp(ControlledGateOperation::PSWAP, all_threading,
+                               all_memory_model, all_qubit_numbers,
+                               KernelType::LM);
     instance.assignKernelForOp(ControlledGateOperation::SingleExcitation,
                                all_threading, all_memory_model,
                                all_qubit_numbers, KernelType::LM);
@@ -336,6 +345,9 @@ void assignKernelsForControlledGeneratorOp_Default() {
                                all_threading, all_memory_model,
                                all_qubit_numbers, KernelType::LM);
     instance.assignKernelForOp(ControlledGeneratorOperation::IsingZZ,
+                               all_threading, all_memory_model,
+                               all_qubit_numbers, KernelType::LM);
+    instance.assignKernelForOp(ControlledGeneratorOperation::PSWAP,
                                all_threading, all_memory_model,
                                all_qubit_numbers, KernelType::LM);
     instance.assignKernelForOp(ControlledGeneratorOperation::SingleExcitation,